Introducing celery utils module instead of using celery root one
